Updated openstack/openstack Project: openstack/horizon  49f8b944e3dc553bd394ba98a69367c2ff157f1b
Adding cluster_template and cluster panels for Sahara

The cluster_template and cluster panels give the ability
to define and launch a hadoop cluster (made from
nodegroup_templates).  They are in the same commit because
some of the functionality of cluster_templates requires
that the cluster panel also exist.
This code was originally from:
https://github.com/openstack/sahara-dashboard

Change-Id: I85534b0dd97ce5def5f3a447d68c6aa206964f0a
Partial-Implements: blueprint merge-sahara-dashboard
Co-Authored-By: Nikita Konovalov <nkonovalov@mirantis.com>
Co-Authored-By: Dmitry Mescheryakov <dmescheryakov@mirantis.com>