Updated openstack/openstack Project: openstack/horizon  8c9b86f1300e3886ba606c9fd569e6698163b20d
Fixes bad performance when editing project members

The bug is caused by consecutive client calls when
fetching users' roles in a project.

The fix was to use the /v3/role_assignments endpoint.
Which retrieves all roles in a single client call
(dashboards/admin/projects/workflows.py:
UpdateProjectGroupsAction and UpdateProjectUsersAction).

The compatibility to the previous keystoneclient version
is maintained.

In a scenario with 1000 users with a role in a project
(using devstack), the time to render the Modify Users
page dropped from 45s to 2.5s

Change-Id: Ic72ebea0db226faf98c5e04d00d76dedd1fb01c1
Closes-bug: #1278920