Updated openstack/openstack Project: openstack-dev/devstack  27eefd87298645b30972a0c0b5a32c3a699521b8
Remove output filter on rabbitmq install

The original workaround was for an init script that grabbed stdout.
This was some time ago and I think it's safe to remove.

It can be a problem on Fedora; out-of-the-box some old F20 images can
have a broken dependency when rabbitmq installs.  With this in place,
all helpful output goes into the temp log file, but it is never shown
because the install_packages function exits the script.  Thus things
just stop and you have no idea why.

Change-Id: I2f5b934492a8c9d7b93e89fdcfa776bf15f25cb8