Updated openstack/openstack Project: openstack/nova  679b92eb7b18a9f5d393fbb8678916e22d8e6b60
Store volume backed snapshot in current tenant.

Fix owner of a creating volume backed snapshot.

Snapshot of an instance booted on a volume based on another tenant's
public image is created in the wrong tenant when invoked by admin.

Snapshot metadata (including owner) is based on image metadata. But
when the snapshot is being created by admin, Glance doesn't change
it's owner if it's set. So we forcibly remove owner (tenant)
attribute from image metadata.

Change-Id: I662dfa4f81e24cb2553ffa2578f4c8530eee9fd3
Closes-Bug: #1322195