Updated openstack/openstack Project: openstack/cinder  7a9f97fe76fe8384d600bee39700ff3c740b592e
Fixes Cinder volume upload to image on windows

Due to fact that when uploading a volume to an image the file is
opened in text mode and the different line endings style on
windows glance client won't be able to upload the desired file.
The solutions is to simply open the file in binary mode when
performing this action.

Change-Id: I51b084ce38e1aaad4ded63b75b097137dedc088c
Closes-bug: 1346911