Updated openstack/openstack Project: openstack/cinder  078f3ea1bf1840a4e9ca3d58a37a432a664f93cc
Implements new 'bootable' option for manage existing volume.

A 'bootable' option (defaults to False) can be passed as part of
the HTTP POST request to manage an existing volume.  This is
passed into the database when the volume is created.

Also changes all drivers to use the more agnostic keys
'source-name' or 'source-id' when specifying
which existing volume to manage.  This change is to support the
cinderclient change in https://review.openstack.org/#/c/76216/
to make it easier for CLI users to manage existing volumes
without having to know the implementation-specifics of the driver.

Change-Id: I58eaa75181fd0b606628584308fd83bdeaa8b64f
Implements: blueprint add-export-import-volumes