Updated openstack/openstack Project: openstack/nova  89e1dd73b8de4628908c10a6cfbbd79adf6e61be
Raise NotImplemented default-security-group-rule api with neutron

Neutron currently does not implement the nova extension
default-security-group-rule which allows an admin to set which rules are in
the default security group when a tenant is onboarded. Implementing this
type of functionality should be done in neutron directly as proxying the
api calls to do this from nova is not possible as neutron creates the
default security group itself. Previously, if one tried to use this api
with neutron a 500 error was raised, now a 501 error is raised.

Change-Id: I6d1bb1c9c2d79278a85b96aea800b4b1f3489225
Closes-bug: 1326958