Updated openstack/openstack Project: openstack/nova  ccb65bf16ec38cc3f61c19c4d920af51c05657c9
Add quota limit create/update methods to Quotas object

This adds two methods to the Quotas object that are effectively
passthroughs to the DB api calls. Since the Quotas object isn't
actually a representation of what is in the database, and because
we've (incorrectly) got some logic in the DB API around how the
quotas are applied, this is just a temporary step. Ideally, we
would (will) refactor Quotas in several ways in the near future,
at which time the way all of this works can be revisited and
cleaned up.

Related to blueprint compute-manager-objects-juno

Change-Id: If93614cd27e56667e0f106e8045dd1adc633ff1b