Updated openstack/openstack Project: openstack/python-novaclient  8d979cfa4b51597c59268ecfd7f051cf1b8cdb94
Description is mandatory parameter when creating Security Group

As per the bug, description should not be a mandatory parameter
as it creates inconsistency with other opestack items which has
description as an optional parameter.

So in the fix, the description argument has been made optional.
The old command was:
nova secgroup-create <name> <description>
Now new command will be:
nova secgroup-create <name> [<description>]

Patch updated based on the comments received

Updated commit message for gerrit autolink

Closes-Bug: #1332133
Change-Id: I2ded915be4a57f3c46e6cf5d1bf00e8c917cd0b9