Updated openstack/openstack Project: openstack/nova  cc417a86c238db3ba4d29a4cad2f0a167701251d
virt: force TCG with libguestfs unless KVM is enabled in libvirt

If the libvirt driver has not been configured to use KVM, then
the libguestfs module should be forced to use TCG. This is
particularly important when running Nova inside a VM, which
might claim to have VMX/SVM support when it is in fact broken.
This will avoid libguestfs hanging in such scenarios.

Resolves-bug: #1286256
Change-Id: I9316dcedd65244c60d468b270311f032b45b051f