Updated openstack/openstack Project: openstack/nova  4b67c5748797567765fdfaa253275c4fcf870962
Stub out rpc notifications in ec2 cloud unit tests

The ec2 cloud unit tests are running through the compute
manager and if something fails and is wrapped with the
wrap_exception decorator, we can get big ugly stack traces
in the log. We should be using the fake_notifier in these
tests like we do in the nova.tests.compute.test_compute
setup.

This also changes the test_ec2_validate test to use
local conductor like the other cloud tests that are going
through conductor.

Change-Id: I59bed0903aeef4086872c10d93a06f34b072dce3
Partial-Bug: #1348661