Updated openstack/openstack Project: openstack/nova  b29772d237cc018c7118d64933e3d7576d197019
Pass errors from detach methods back to api proc

The initial implementation of the detach_interface method was built as
cast invocations.  The respective attach method was built as a call
invocation.  What this led to was, if the user attempted to attach an
incorrect port, that error would be passed back up to the API
properly.

However, if the user attempted to detach an invalid interface, the
response from the API would be an HTTP 202 (because the cast assumes
success).  The root error may be logged in the respective nova compute
log, but it did not make its way back up to the user.

This change updates the decorators to the manager methods to update
the instance's fault when the attach or detach encounters an issue.

Change-Id: I111b1c8a492f12587f86ab83825060029605390d
Closes-Bug: #1339098