Updated openstack/openstack Project: openstack/python-openstackclient  01f7a5f9f59705deb5c800b78457f69c1e73cdbd
Change V2 image tests to actually run V2 image code

The current tests for image do not run v2 image code, changing
that portion also made it's only test fail.
I opted to change the image delete code and not the test, since
passing the object ID is more in line with the rest of the
project code.

Change-Id: I62e13c063a5d68279dbbf31e59266db6285d73bf