Updated openstack/openstack Project: openstack/neutron  7c1ee05c7e815ee9aefdd7a0d85855f3c70620c2
Set python hash seed to 0 in tox.ini

New tox (>=1.7.0) sets a random python hash seed by default. This is
generally good for testing because it will help keep projects working
regardless of the hash seed, but neutron unittests don't currently pass
with a random hash seed so set it to the python default seed.

This change will allow us to use new tox again and remove the
restriction on tox<=1.6.1 to run unittests.

Note this change will need to be backported to the stable branches to
keep unittests there working with new tox as well.

Change-Id: If151efeb8caae3111515191e238d6e6ae491cd44