Updated openstack/openstack Project: openstack/tempest  02ac93e49800310f8cd1cc46615a2b11a363f54b
Skip migrate server invalid state if resize is off

The migrate (aka cold migration) can be considered a special case of the
resize with a flavor_id equals to None. For that reason to perform
the migrate tests we need to be sure that the resize is enabled.
This patch adds a skip decorators to verify that.

Change-Id: I3a02b1bd210e3934dfd61d4182b301cc33d8a962