Updated openstack/openstack Project: openstack/cinder  a61713fbdbdfaefe43e4efdfa89a04ad70c8c5d5
Remove cinder-clear-rabbit-queues

cinder-clear-rabbit-queues should be removed because it does not
appear to be widely used and there is no user documentation for
it. Also, RabbitMQ now includes a management plugin and CLI,
rabbitmqadmin, that allows you to delete queues and exchanges.
This script was also removed in nova as part of
https://review.openstack.org/#/c/40355/.

DocImpact: The RabbitMQ management plugin ships with rabbitmqadmin.
It supports RabbitMQ version 2.8.x and above.
Closes-Bug: #1348859

Change-Id: I77bb54065280771e5cc64835af9c584626ecf8b8