Updated openstack/openstack Project: openstack/cinder  0058f5ac5b58b91091fbc23e847ce0a4dd54141e
Get updated model info on volume transfer

On some drivers CHAP credentials are tied to accounts
which may be unique for each OS and/or backend Tenant.

For example:
  Volume is created with project_id xyz
  Backend creates an account for xyz with specific CHAP

  Volume is transferred to project_id abc

Currently we don't update the model_info, so even though
on the backend we updated the owner and CHAP settings we
never propogated that back up to the DB object so the volume
can't be used by the new owner.

This patch just adds an option to return model_update
on the accept_transfer call to the driver and updates
the db accordingly.

Also adds a call in the SF driver to actually get the
new model info to be fed back to the manager.

Closes-Bug: #1349475

Conflicts:
	cinder/volume/drivers/solidfire.py

Change-Id: Ie0447cdad69c4fbee99b4b6b1d3cacdfdd14137d
(cherry picked from commit 09f6fe636416ac63e1463d14d469af8f93a6ddf0)