Updated openstack/openstack Project: openstack/nova  fb63d85a0403c125dd187e7fdb787fc9f074b8fe
Scheduler: throw exception if no configured affinity filter

If the scheduler hint indicates that the scheduling should perform
either anti-affinity or affinity scheduling and the relevant filter
is not configured then a NoValidHost exception will be thrown.

This is valuable if an existing OpenStack installation is running
and these filters are not defined after an upgrade.

Change-Id: I79bb44ad7481b3ff924687a8d6afdd6c715c0b59
Closes-bug: #1302238