Updated openstack/openstack Project: openstack/swift  7b0b949bf8b00d0349c2182004556c9a69293a4c
Pass disk usage options when using --all

This commit allows the arguments --human-readable and --top to be
used with --all to customize the output of the disk usage portion.

There are no unit tests attached to this commit, I have created and
referenced bug #1344200 which shows there is a larger bug to address
to create several missing tests.

Change-Id: I46b8359533989efc3067971acaafec1d0cbc2b9f
Closes-Bug: #1344178
Related-Bug: #1344200