Updated openstack/openstack Project: openstack/tempest  14aaff9da4c63e85519b63a9f19a93f02bc9c5fd
Move CommandFailed exception to tempest.exceptions

When the CommandFailed exception was added it was only being used by
the CLI tests. However, it is generally useful for anything that is
using subprocess to make external calls. This patch moves it to
tempest.exceptions to make using it simpler for non-cli tests to use
the exception.

Change-Id: Ibf5f1cbbb847d32976b54c4484acfc3c0e3b4f48