Updated openstack/openstack Project: openstack/ceilometer  c638d043b1845c1c6e7b0e8b50f921494d9cdf5c
Add cumulative and gauge to aggregator transformer

Currently, the aggregator transformer just sums up the volumes of all
aggregated samples.

This patch improves it so it returns just the latest volume for cumulative
meters and the average of volumes for gauge meters.

DocImpact: Aggregator transformer now works for gauge and cumulative as well
Change-Id: I172a7298dc063ff36b2a209df25d6650debb5a14
Closes-bug: #1342668