Updated openstack/openstack Project: openstack/cinder  28647b4ecdf07088c472fd75539327d1c6d1fc56
Add hacking check for vim headers

We have a number of commits that have removed the vim
headers from openstack/common components as well as
commits to remove the headers from drivers.  We, however,
haven't put a hacking check in to make sure that they don't
sneak back into the code as is evidenced by the fact that some
files failed the check.

This commit adds in the check, the associated test cases and
fixes the couple of files that had a vim header again.

Change-Id: I727bd2aff7ac7b909c9e7cd4f3639f5873793e97