Updated openstack/openstack Project: openstack/keystone  a19dd6c2ab18d77c49167fccc82dda69387d5b8e
Update middleware that was moved to keystonemiddleware

S3Token middleware has been moved (and slated for removal) in the
Juno Release. This Middleware has been removed from Keystone.

Keystone now uses the middleware located in keystonemiddleware for
providing the code for ec2_token middleware. The middleware class
in Keystone has been deprecated and is slated for removal in the
L release.

Change-Id: Iecda3400c0450bbeed42093dc916825d89b2c3fe
bp: split-middleware