Updated openstack/openstack Project: openstack/horizon  5696c0da700444ba70a36a5c44722694d4e3bfad
Updated the steps of the tutorial

The existing tutorial is outdated, hard to follow and does not
work.

Update or rewrite the steps of the tutorial to create a dashboard,
a panel, a tab group, a tab and a table which contains some data
from backend a table which contains some data from backend.

Make sure the steps work.

Updated the brief explanation as well.

Closes-Bug: #1298490

Change-Id: I7c6769fa6b8cd5488afc03e8a284342549a6745f