Updated openstack/openstack Project: openstack/nova  4903a9837be228e01eabb71748b7fdba671ab7cd
Fix Resource tracker should report virt driver stats

If the virt driver provides any data for resource stats it is
lost whenever the resource tracker updates its own view of stats.
Moreover, if the resource tracker has not instances to track it
only reports the driver's view, which might be nothing.

This fix adds the driver's view of stats to the resource tracker
stats to make sure they are correctly handled.

Change-Id: Icb19148660bca542a8120ecab064551d67ac28af
Closes-bug: #1348288