Updated openstack/openstack Project: openstack/sahara  6159492733c384749f4afda6a00055bd39c65212
Set python hash seed to 0 in tox.ini

New tox (>=1.7.0) sets a random python hash seed by default. This is
generally good for testing because it will help keep projects working
regardless of the hash seed, but nova unittests don't currently pass
with a random hash seed so set it to the python default seed.

This change will allow us to use new tox again and remove the
restriction on tox<=1.6.1 to run unittests.

Redundant setenvs for specific test envs are removed to prevent them
from overriding the PYTHONHASHSEED value.

Note this change will need to be backported to the stable branches to
keep unittests there working with new tox as well.

Change-Id: I7e2db17668bf522eb9ff4d77eb1ba47be39b52fb
Co-Authored-By: Clark Boylan <clark.boylan@gmail.com>