Updated openstack/openstack Project: openstack/heat  3db5aa549530e3098dee06a4e7b3b3c2e5e47b83
Allow Resources to select paths from attributes

Previously when get_attr included a path to select from the attribute, the
selection was done by the get_attr intrinsic function. Starting with the
Juno version of HOT, move control of this to the resource plugin, so that
individual resources (particularly ResourceGroup) can handle it themselves.

Related-Bug: #1341048
Change-Id: Ic3dc94a99a19e28a4f775198ff6e960e2efb8256