Updated openstack/openstack Project: openstack/swift  ac74b85f3e55f3b01b1d9fb01a4541730bc4e2a5
Put proxy-logging in object expirer's sample config

It's generally better to have logs for something than to not have
logs. This way, the object expirer (if using the sample config as a
starting point) will log what it does.

Note that the container reconciler's sample config already contains
proxy-logging, as does the proxy server's. The object expirer is the
odd man out.

Change-Id: I32aac99131746501820319b94405440c1934a694